+++ pkg/payload/payload.go
@@ -19,7 +19,7 @@ type B2CRequest struct {
 	CommandID          string  `json:"CommandID,omitempty"`
 	Amount             float64 `json:"Amount,omitempty"`
 	PartyA             int32   `json:"PartyA,omitempty"`
-	PartyB             int32   `json:"PartyB,omitempty"`
+	PartyB             int64   `json:"PartyB,omitempty"`
 	Remarks            string  `json:"Remarks,omitempty"`
 	QueueTimeOutURL    string  `json:"QueueTimeOutURL,omitempty"`
 	ResultURL          string  `json:"ResultURL,omitempty"`
@@ -30,7 +30,7 @@ type B2CRequest struct {
 type ReversalRequest struct {
 	CommandID              string `json:"CommandID,omitempty"`
 	ReceiverParty          int32  `json:"ReceiverParty,omitempty"`
-	ReceiverIdentifierType int32  `json:"ReceiverIdentifierType,omitempty"`
+	ReceiverIdentifierType int64  `json:"ReceiverIdentifierType,omitempty"`
 	Remarks                string `json:"Remarks,omitempty"`
 	Initiator              string `json:"Initiator,omitempty"`
 	SecurityCredential     string `json:"SecurityCredential,omitempty"`
